grml plugin now uses own directory (not "live"

+++ plugins/grml.sh
@@ -2,9 +2,8 @@
 set -e
 . ./functions.sh
 #GRML plugin for multicd.sh
-#version 6.3
-#Copyright (c) 2011 libertyernie
-#modified quick and dirty by T.Ma.X. N060d9 to work with grml
+#version 6.6
+#Copyright (c) 2011 libertyernie, T.Ma.X. N060d9
 #Permission is hereby granted, free of charge, to any person obtaining a copy
 #of this software and associated documentation files (the "Software"), to deal
 #in the Software without restriction, including without limitation the rights
@@ -32,34 +31,38 @@ elif [ $1 = copy ];then
 	if [ -f grml.iso ];then
 		echo "Copying GRML..."
 		mcdmount grml
-		mkdir $WORK/live
+		mkdir $WORK/grml
 		mkdir $WORK/conf
 		mkdir $WORK/boot/grml
-		cp -r $MNT/grml/live/* $WORK/live #Compressed filesystem.
+		cp -r $MNT/grml/live/* $WORK/grml #Compressed filesystem.
 		cp $MNT/grml/boot/grml/linux26 $WORK/boot/grml/ #Kernel. See above.
 		cp $MNT/grml/boot/grml/initrd.gz $WORK/boot/grml/initrd.gz #Initial ramdisk. See above.
 		cp $MNT/grml/GRML/grml-version $WORK/boot/grml/grml-version
 		cp $MNT/grml/conf/bootid.txt $WORK/conf/ #needed for booting
 #getting  files into one and with write access:
-		cp $MNT/grml/boot/isolinux/default.cfg $WORK/boot/grml/grml.cfg #isolinux menufile
-		cat $MNT/grml/boot/isolinux/grml.cfg >> $WORK/boot/grml/grml.cfg #isolinux menufile
+		cp $MNT/grml/boot/isolinux/default.cfg $WORK/boot/grml/grml.cfg #isolinux menufile - temporary
+		cat $MNT/grml/boot/isolinux/grml.cfg >> $WORK/boot/grml/grml.cfg #adding to menufile
 		umcdmount grml
 	fi
 elif [ $1 = writecfg ];then
 if [ -f grml.iso ];then
-VERSION=$(awk '{print substr($2,1,7)}' $WORK/boot/grml/grml-version) #Getting grml Version
-echo "menu begin --> GRML $VERSION
-" >> $WORK/boot/isolinux/isolinux.cfg
+	VERSION=$(awk '{print substr($2,1,7)}' $WORK/boot/grml/grml-version) #Getting grml Version
+	echo "menu begin --> GRML $VERSION
+	" >> $WORK/boot/isolinux/isolinux.cfg
 
-sed -i".bak" '1d' $WORK/boot/grml/grml.cfg #deleting some lines we don't want
-sed -i".bak" '2d' $WORK/boot/grml/grml.cfg 
-sed -i".bak" '/menu end/ i\label back\n   menu label Back to main menu...\n   com32 menu.c32\n'  $WORK/boot/grml/grml.cfg #insert back to main menu
-rm $WORK/boot/grml/grml.cfg.bak #bak file from sed not needed
+	sed -i".bak" '1d' $WORK/boot/grml/grml.cfg #deleting some lines we don't want
+	sed -i".bak" '2d' $WORK/boot/grml/grml.cfg
+	sed -i".bak" '/menu end/ i\label back\n   menu label Back to main menu...\n   com32 menu.c32\n'  $WORK/boot/grml/grml.cfg #insert back to main menu
 
-cat $WORK/boot/grml/grml.cfg >>  $WORK/boot/isolinux/isolinux.cfg #putting everything together
+	sed -i -e 's^boot=live^boot=live live-media-path=/grml^g' $WORK/boot/grml/grml.cfg
 
-rm $WORK/boot/grml/grml.cfg #not needed any longer
+	if [ -f $WORK/boot/grml/grml.cfg.bak ];then
+		rm $WORK/boot/grml/grml.cfg.bak #bak file from sed not needed
+	fi
 
+	cat $WORK/boot/grml/grml.cfg >> $WORK/boot/isolinux/isolinux.cfg #putting everything together
+
+	rm $WORK/boot/grml/grml.cfg #not needed any longer
 fi
 else
 	echo "Usage: $0 {links|scan|copy|writecfg}"
